fix: address all CodeRabbit review findings
- Fix mouse button event flags to match cef_event_flags_t (1<<4, 1<<5, 1<<6)
- Make runtime.Init thread-safe with sync.Once
- Add nil guard in browser.Close
- Use int64 intermediate for BufferSize to prevent 32-bit overflow
- Add defensive nil check in refcount.Init
- Use Dlsym before RegisterFunc in loader to avoid panics
- Add named constant for cef_version_info chrome major entry
- Replace os.Unsetenv with t.Setenv for test isolation
- Fix hardcoded path in parser test, use CEF_DIR env var
- Fix pointer detection in parser mapType
- Remove unused tag param from parseStruct
- Remove unused lowerFirst template function
- Relax config validation (version has a default)
- Add negative validation test cases
- Add non-nil LifeSpanHandler builder test
- Improve refcount test error message
- Document Linux-only scope and Dlclose rationale

+++ cef/client.go
@@ -87,7 +87,7 @@ func newClientState(c *Client) *clientState {
 				Width:      width,
 				Height:     height,
 				Buffer:     buffer,
-				BufferSize: int(width) * int(height) * 4, // BGRA
+				BufferSize: int(int64(width) * int64(height) * 4), // BGRA
 				DirtyRects: rects,
 			}
 			c.renderHandler.OnPaint(evt)
